Connect getAllAssociations and move client logic from page
