+++ src/app/dashboard/(active)/web/associations/types.ts
@@ -1,7 +1,7 @@
 export type Association = {
   id: number
   name: string
-  icon: string
   descriptionIt: string
   descriptionEn: string
+  logoSvg: string | null
 }
